fix(extract): apply Darwin is_extern_c fix to SemanticIR locality too; split FIFO tests out
Codex review on PR #1048 found two real issues:

1. extract/manifest_semantic_ir.py's _is_locally_linked_function/
   _is_locally_linked_variable duplicate tu_merge.py's own locality
   logic (documented as intentional -- extract/ may not import the
   root-level tu_merge module per ADR-061) but still used the same
   broken mangled == name check the previous commit fixed only in
   tu_merge.py. Apply the identical Darwin-aware is_extern_c fix here
   too, so the persisted SemanticIR occurrences agree with the flat
   MergedTuFragments.functions/.variables lists instead of silently
   re-collapsing what tu_merge.py now correctly keeps distinct.

2. Adding the FIFO write helper to tests/test_snapshot_compression.py
   grew it past its architecture/debt.yaml no-growth baseline (1499
   lines). Split the two FIFO write-through tests and their shared
   _write_through_fifo() helper into a new sibling leaf module,
   tests/test_snapshot_compression_fifo.py, following the same
   established pattern as this file's existing
   test_snapshot_compression_skippable_frames.py/
   test_snapshot_compression_public_api_scale.py siblings.

Verified: mypy clean, ruff clean, python scripts/check_architecture.py
reports 0 errors (previously 1 debt-no-growth error), targeted tests
pass (178 passed), plus a direct primitive-level check that
_is_locally_linked_function/_is_locally_linked_variable now agree with
tu_merge.py's _function_key/_variable_key on the exact Darwin-quirk
shapes.
